DCI/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
djangorestframework==3.14.0
filelock==3.13.1
frozendict==2.4.0
fsspec==2023.12.2
html5lib==1.1
idna==3.6
Jinja2==3.1.2
lxml==5.1.0
MarkupSafe==2.1.3
mpmath==1.3.0
multitasking==0.0.11
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
pandas==2.0.3
peewee==3.17.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
torch==2.1.2
triton==2.1.0
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
webencodings==0.5.1
yfinance==0.2.35
